// app/model/PackageInfo.ts
export interface PackageInfo {
    packageId: string;
    flowId: string;
    flowName: string;
    weight: number;
    height: number;
    width: number;
    number: number;
    progress: number;
    orderId: string;
    companyName: string;
    orderedDate: Date;
    deliveryDate: Date;
    orderNumber: number;
    orderYear: number;
    locatioName: string;
    locationId: string;
    length: number;
}